% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{baseline}
\alias{baseline}
\title{Computes the statistical baseline}
\usage{
baseline(labels, iterations = 50)
}
\arguments{
\item{labels}{Vector of all labels contained in the dataset}

\item{iterations}{Number of baseline iterations}
}
\value{
Returns mean of iteration
}
\description{
Computes a statistical baseline running the algorithm multiple times and
calculating the mean of those runs.
}

