% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{distance}
\alias{distance}
\title{Distance function.}
\usage{
distance(data, distmethod = "euclidean", nominals = NULL)
}
\arguments{
\item{data}{Data}

\item{distanceMethod}{Distance method}

\item{iterations}{Number of iterations}

\item{nominal}{Nominal attributes}

\item{baselineIterations}{Number of baseline iterations}
}
\value{
Returns distance matrix.
}
\description{
Expects dataframe \emph{without} class label column
}
\details{
For euclidean and manhattan distances, it uses the `dist` function from base and for HEOM,
it uses the code found in `heom.R`-file.
}

