% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heom.R
\name{heom}
\alias{heom}
\title{An implementation of HEOM metric.}
\usage{
heom(data)
}
\arguments{
\item{data}{The data to calculate the distance matrix from}
}
\value{
Returns a distance matrix
}
\description{
This creates distance matrix from given data.
}
\details{
TODO: This breaks, if the data has only one column. Also, check if the sapply-
way to check numerics and factors is correct, since if there's only one column,
it returns logical vector for all cases, not for the column.

For further reading, see [1].

[1] Wilson, D. Randall, and Tony R. Martinez. 1997. “Improved Heterogeneous
Distance Functions.” Journal of Artificial Intelligence Research 6: 1–34.
doi:10.1613/jair.346.
}
