% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{maxChanges}
\alias{maxChanges}
\title{Calculate theoretical maximum value for label changes.}
\usage{
maxChanges(labels, current = NULL)
}
\description{
Calculate theoretical maximum value for label changes.
}

