% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{n_changes}
\alias{n_changes}
\title{Calculate the number of label changes}
\usage{
n_changes(collection_vector)
}
\arguments{
\item{collection_vector}{A collection vector produced by \code{traverse}.}
}
\value{
The number of changes in the collection vector
}
\description{
If current and next label are not the same, then counter is incremented by
one.
}
