% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{run}
\alias{run}
\title{This is the main function, which groups all smaller parts together and returns
the final results.}
\usage{
run(data, classlabel = NULL, distanceMethod = "euclidean",
  usecase = "single", iterations = 10, baselineIterations = 50,
  classes = NULL, columns = NULL, nominals = NULL)
}
\arguments{
\item{data}{A data frame to process}

\item{classlabel}{Column name or index for class label}

\item{distanceMethod}{Distance method to use; must be one of following: euclidean, 
manhattan or heom.}

\item{usecase}{A usecase to select; must be one of following: single, classes, variables 
or all}

\item{baselineIterations}{A number of iteration for baseline calculation}

\item{classes}{List of class included in the calculation}

\item{columns}{List of indices or names of columns included in the calculation}

\item{nominals}{List of indices or names of those columns that are nominal; not in use at
the moment}

\item{iteration}{A number of iterations for Scatter-algorithm}
}
\value{
result
}
\description{
Takes a dataframe as input. Note that the dataframe is assumed to be
normalized to the range [0, 1].
}
\details{
The classlabel is the name or index of the column containing the classlabel
and it is moved to the last position and ignored when calculating distances
etc.
}

