\name{Scatter-R-package}
\alias{Scatter-R-package}
\alias{Scatter-R}
\docType{package}
\title{
\packageTitle{Scatter-R}
}
\description{
\packageDescription{Scatter-R}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Scatter-R}
\packageIndices{Scatter-R}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Scatter-R}

Maintainer: \packageMaintainer{Scatter-R}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
