% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter-preprocess.R
\name{scatter.preprocess}
\alias{scatter.preprocess}
\title{Transform a dataframe into a format suitable for the Scatter algorithm}
\usage{
scatter.preprocess(df, classvar, included.attributes = NULL,
  included.classes = NULL, binarized = NULL, scaled = NULL,
  na.action = NULL)
}
\arguments{
\item{df}{A dataframe to be preprocessed for use with the Scatter algorithm}

\item{classvar}{Name or index of the column containing the class label}

\item{included.attributes}{Names or indices of columns/attributes to be included in the processed dataframe (columns that are not in included attributes are stripped; if not given, by default all are included)}

\item{included.classes}{Names or indices of classes to be included in the processed dataframe (rows where class label is not in included classes are stripped; if not given, by default all are included)}

\item{binarized}{Names or indices of nominal attributes/columns to be binarized in the preprocessing}

\item{scaled}{Names or indices of numeric attributes/columns to be scaled in the preprocessing}

\item{na.action}{Action for missing values (class/column/rmrows/nothing/NULL)}
}
\value{
A dataframe preprocessed according to the arguments
}
\description{
Transform a dataframe into a format suitable for the Scatter algorithm
}

