\name{ScatterR-package}
\alias{ScatterR-package}
\alias{ScatterR}
\docType{package}
\title{
\packageTitle{ScatterR}
}
\description{
\packageDescription{ScatterR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ScatterR}
\packageIndices{ScatterR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ScatterR}

Maintainer: \packageMaintainer{ScatterR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
