% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{spower}
\alias{spower}
\title{Calculates separation power}
\usage{
spower(z, s)
}
\arguments{
\item{z}{Baseline}

\item{s}{Raw Scatter value}
}
\value{
Returns the difference between baseline and raw scatter value.
}
\description{
Calculates separation power
}
