% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{usecase.all}
\alias{usecase.all}
\title{Usecase `all` runs the \code{classes} for all variables separately}
\usage{
usecase.all(data, distanceMethod = "euclidean", iterations = 10,
  nominal = c(), baselineIterations = 50)
}
\description{
Usecase `all` runs the \code{classes} for all variables separately
}

