% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{usecase.class}
\alias{usecase.class}
\title{Usecase \code{class} runs the algorithm for each class separately.}
\usage{
usecase.class(data, distanceMatrix, iterations = 10, nominal = c(),
  baselineIterations = 50)
}
\arguments{
\item{data}{Data}

\item{iterations}{Number of iterations}

\item{nominal}{Nominal attributes}

\item{baselineIterations}{Number of baseline iterations}

\item{distanceMethod}{Distance method}
}
\value{
TBD
}
\description{
Usecase \code{class} runs the algorithm for each class separately.
}
\examples{
TBD
}

