% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{usecase.variable}
\alias{usecase.variable}
\title{Usecase \code{variable} runs the algorithm for each variable separately.}
\usage{
usecase.variable(data, distance_method = "euclidean", iterations = 10,
  nominal = c(), baseline_iterations = 50, quiet = FALSE)
}
\arguments{
\item{data}{Data}

\item{distance_method}{Distance method}

\item{iterations}{Number of iterations}

\item{nominal}{Nominal attributes}

\item{baseline_iterations}{Number of baseline iterations}

\item{quiet}{A flag that controls whether messages are printed or not}
}
\value{
TBD
}
\description{
Usecase \code{variable} runs the algorithm for each variable separately.
}
\examples{
#TBD
}
