% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{distance}
\alias{distance}
\title{Distance function.}
\usage{
distance(data, distanceMethod = "euclidean", nominals = NULL)
}
\arguments{
\item{data}{The data}

\item{distanceMethod}{Distance method; must be one of following:
\code{euclidean}, \code{manhattan} or \code{heom}}

\item{nominals}{Nominal attributes; not in use at the moment}
}
\value{
Returns a distance matrix.
}
\description{
For euclidean and manhattan distances, it uses the `dist` function from
base and for HEOM, it uses the code found in `heom.R`-file. Note, that 
this function expects data to be without class-column!
}

