% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Calculate raw Scatter value}
\usage{
scatter(labels)
}
\arguments{
\item{labels}{The set of labels found from data}
}
\value{
Returns a raw Scatter value
}
\description{
Calculate raw Scatter value
}

