% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{traverse}
\alias{traverse}
\title{Generate collection vector}
\usage{
traverse(labels, distance_matrix, seed = FALSE)
}
\arguments{
\item{labels}{The class labels}

\item{distance_matrix}{The distance matrix calculated from \code{data}}

\item{seed}{Seed for randomzing starting point}
}
\value{
Returns the vector of labels
}
\description{
Traverse the dataset using nearest neighbour method, recording label changes
as we go.
}
