% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{usecase.all}
\alias{usecase.all}
\title{Usecase `all` runs the \code{classes} for all variables separately}
\usage{
usecase.all(data, distance_method = "euclidean", iterations = 10,
  baseline_iterations = 50)
}
\arguments{
\item{data}{Data}

\item{distance_method}{Distance method}

\item{iterations}{Number of iterations}

\item{baseline_iterations}{Number of baseline iterations}
}
\value{
TBD
}
\description{
Usecase `all` runs the \code{classes} for all variables separately
}
