% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{usecase.all}
\alias{usecase.all}
\title{Usecase `all` runs the \code{classes} for all variables separately}
\usage{
usecase.all(data, distanceMethod = "euclidean", iterations = 10,
  nominal = c(), baselineIterations = 50, quiet = FALSE)
}
\arguments{
\item{data}{Data}

\item{distanceMethod}{Distance method}

\item{iterations}{Number of iterations}

\item{nominal}{Nominal attributes}

\item{baselineIterations}{Number of baseline iterations}

\item{quiet}{A flag that controls whether messages are printed or not}
}
\value{
TBD
}
\description{
Usecase `all` runs the \code{classes} for all variables separately
}

