% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{usecase.class}
\alias{usecase.class}
\title{Usecase \code{class} runs the algorithm for each class separately.}
\usage{
usecase.class(data, distanceMatrix, iterations = 10, nominal = c(),
  baselineIterations = 50, quiet = FALSE)
}
\arguments{
\item{data}{Data}

\item{distanceMatrix}{A distance matrix. This is different than other
usecases as it doesn't take \code{distanceMethod} but 
\code{distanceMatrix}.}

\item{iterations}{Number of iterations}

\item{nominal}{Nominal attributes}

\item{baselineIterations}{Number of baseline iterations}

\item{quiet}{A flag that controls whether messages are printed or not}
}
\value{
TBD
}
\description{
Usecase \code{class} runs the algorithm for each class separately.
}
\examples{
#TBD
}

