% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{usecase.class}
\alias{usecase.class}
\title{Usecase \code{class} runs the algorithm for each class separately.}
\usage{
usecase.class(data, distance_matrix, iterations = 10,
  baseline_iterations = 50)
}
\arguments{
\item{data}{Data}

\item{distance_matrix}{A distance matrix. This is different than other
usecases as it doesn't take \code{distance_method} but
\code{distance_matrix}. This is because the usecase.class is reused
in usecase.all.}

\item{iterations}{Number of iterations}

\item{baseline_iterations}{Number of baseline iterations}

\item{A}{flag that controls whether messages are printed or not}
}
\value{
TBD
}
\description{
Usecase \code{class} runs the algorithm for each class separately.
}
\examples{
#TBD
}
