% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{usecase.single}
\alias{usecase.single}
\title{Usecase `single`. This runs scatter algorithm for the whole dataset.}
\usage{
usecase.single(data, distance_method = "euclidean", iterations = 10,
  nominal = c(), baseline_iterations = 50, quiet = FALSE)
}
\arguments{
\item{data}{Data}

\item{distance_method}{Distance method}

\item{iterations}{Number of iterations}

\item{nominal}{Nominal attributes}

\item{baseline_iterations}{Number of baseline iterations}

\item{quiet}{A flag that controls whether messages are printed or not}
}
\value{
TBD
}
\description{
Usecase `single`. This runs scatter algorithm for the whole dataset.
}
\examples{
#TBD
}
