% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{usecase.single}
\alias{usecase.single}
\title{Usecase `single`. This runs scatter algorithm for the whole dataset.}
\usage{
usecase.single(data, distanceMethod = "euclidean", iterations = 10,
  nominal = c(), baselineIterations = 50, quiet = FALSE)
}
\arguments{
\item{data}{Data}

\item{distanceMethod}{Distance method}

\item{iterations}{Number of iterations}

\item{nominal}{Nominal attributes}

\item{baselineIterations}{Number of baseline iterations}
}
\value{
TBD
}
\description{
Usecase `single`. This runs scatter algorithm for the whole dataset.
}
\examples{
#TBD
}

